support lcov files without TN
Some lcov files don't have a `TN:` field.
The parser currently crashes in this case.
Just set `current_test_name` to empty at the beginning
to handle this case.

The documentation
https://manpages.ubuntu.com/manpages/noble/man1/geninfo.1.html
seems to say that the testname is indeed optional:

```
If available, a tracefile begins with the testname which is stored in the following format:

         TN:<test name>
```